api/pyproject.toml
[project]
name = "rainref-api"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "pydantic>=2.6",
  "sqlalchemy>=2.0",
  "pgvector>=0.3.5",
  "psycopg[binary,pool]>=3.2",
  "alembic",
  "python-multipart",
  "azure-storage-blob",
  "httpx",
  "tenacity",
  "opentelemetry-sdk",
  "opentelemetry-instrumentation-fastapi",
  "opentelemetry-exporter-otlp",
  "langchain",
  "langgraph==0.2.24",
  "scikit-learn",
  "numpy",
  "jinja2",
  "pypdf",
  "python-docx",
  "rapidfuzz",
  "uvloop; platform_system != 'Windows'"
]

[tool.ruff]
line-length = 100

[tool.black]
line-length = 100
